refactor(registry): 拆分为 registry/ 子模块（model/builtinSources/lookup/paging）
- model.ts: SourceKind/NewsSource 类型、分组元信息与分页常量
- builtinSources.ts: 内置 SOURCES 数据与网易/公众号构造器
- lookup.ts: findSource/userAgentFor/自定义源 id/公众号合集链接归一
- paging.ts: 分页策略判定、offset 翻页 URL 构造与财联社签名
- registry.ts 保留原路径作为 barrel（vite.config/functions 显式 .ts 导入不受影响）
